[REFACTOR] Split scss variables into smaller files and move component styles into folder
