+++ src/client/react/scss/admin/_dashboard.scss
@@ -1,4 +1,4 @@
-@import '../variables';
+@import '../variables/bp';
 
 #admin-dashboard {
 	@media (min-width: $bp-s) {
