+++ src/client/react/scss/components/_base.scss
@@ -1,4 +1,4 @@
-@import 'variables';
+@import '../variables/bp';
 
 .hero {
 	#base {
