+++ src/client/react/scss/components/_dashboard.scss
@@ -1,4 +1,5 @@
-@import 'variables';
+@import '../variables/bp';
+@import '../variables/colours';
 
 #dashboard {
 	.dashboard {
