+++ src/client/react/scss/components/_description.scss
@@ -1,4 +1,5 @@
-@import 'variables';
+@import '../variables/bp';
+@import '../variables/colours';
 
 .event-description {
 	width: 100%;
