+++ src/client/react/scss/components/_hero.scss
@@ -1,4 +1,6 @@
-@import 'variables';
+@import '../variables/bp';
+@import '../variables/colours';
+@import '../variables/fonts';
 
 $left-margin: 3rem;
 $left-margin-s: 1rem;
