+++ src/client/react/scss/components/_registration-closed.scss
