+++ src/client/react/scss/main.scss
@@ -9,7 +9,7 @@ $icon-font-path: "~@blueprintjs/core/resources/icons";
 // Import animate.css
 @import '../../../../node_modules/animate.css/animate.min.css';
 
-@import 'variables';
+@import 'variables/fonts';
 @import 'fonts';
 
 html, body {
