+++ src/client/react/scss/variables/_bp.scss
@@ -1,0 +1,8 @@
+// Breakpoints
+$bp-xxs: 359px;
+$bp-xs: 480px;
+$bp-s: 640px;
+$bp-m: 768px; // iPad portrait
+$bp-l: 1024px; // iPad landscape
+$bp-xl: 1244px;
+$bp-xxl: 1410px;
