+++ src/client/react/scss/variables/_colours.scss
@@ -1,0 +1,3 @@
+// Colours
+$theme-orange: #E6AA00;
+$theme-orange-light: #FCD45F;
