+++ src/client/react/scss/variables/_fonts.scss
@@ -1,0 +1,4 @@
+// Fonts
+$fallback-font: sans-serif;
+$brand-font: 'Coture', $fallback-font;
+$theme-font: 'Encode Sans', $fallback-font;
