+++ src/client/react/stages/Registration/Register.js
@@ -3,7 +3,7 @@ import RegisterForm from './RegisterForm';
 import Title from '../components/Title';
 import ScrollAnimation from 'react-animate-on-scroll';
 
-import '../../scss/_register.scss';
+import '../../scss/components/_register.scss';
 
 
 class Register extends React.Component {
