+++ src/client/react/stages/RegistrationClosed/Home.js
@@ -6,7 +6,7 @@ import Title from '../components/Title';
 import Description2 from '../components/Description2';
 import ScrollAnimation from 'react-animate-on-scroll';
 
-import '../../scss/_registration-closed.scss';
+import '../../scss/components/_registration-closed.scss';
 
 
 const mapStateToProps = (state, ownProps) => {
